test with missings
